==> ipms/R/00_Main.R <==
# Master IPM sourcing script

# See each sub-script for notes on specifics

# source('ipms/R/01_Climate_Data_Download.R')
source('ipms/R/02_Utility_Functions_and_Dependencies.R')
# source('ipms/R/03_Data_Infiling_and_Preprocessing.R')
# source("ipms/R/04_Data_Exploration.R")
# source('ipms/R/05_Fecundity_Models.R')
# source('ipms/R/06_Surv_Grow_Models.R')
# source('ipms/R/07_Preliminary_Plots.R')
source('ipms/R/08_IPM_Construction.R')
source('ipms/R/09_Lambda_Plots.R')
source('ipms/R/10_Simulated_Parameter_Sensitivity.R')
source('ipms/R/11_LTRE.R')
